♻️ (color range settings): Refactored functions and created new utils js file
