+++ src/stores.js
@@ -28,9 +28,6 @@ export const balloonRangeColor1 = writable('#ff0000');
 export const balloonRangeColor2 = writable('#00ff00');
 export const balloonRangeColorDefinition = writable(10);
 export const balloonInterpolatedColors = writable([]);
-export function updateInterpolatedColors(colors) {
-    balloonInterpolatedColors.set(colors);
-  }
 export const balloonColor = writable('#ff0000');
 export const gameBackgroundColor = writable('#add8e6');
 export const enableCustomLetter = writable(false);
